Moving api tests to separate directory.
