+++ express-backend-api/test/api/session.api.test.ts
@@ -1,15 +1,15 @@
-import {container} from "../config/container.config";
-import {TYPES} from "../config/types.config";
+import {container} from "../../config/container.config";
+import {TYPES} from "../../config/types.config";
 import request from "supertest";
-import {API_URI_LIR} from "../config/api.config"
+import {API_URI_LIR} from "../../config/api.config"
 import {faker} from "@faker-js/faker";
-import app from "../app";
-import {UserPartial} from "../model/user.partial";
-import {UserRepository} from "../repository/user.repository";
-import {User} from "../model/user";
-import {SessionPartial} from "../model/session.partial";
-import {Session} from "../model/session";
-import {SessionRepository} from "../repository/session.repository";
+import app from "../../app";
+import {UserPartial} from "../../model/user.partial";
+import {UserRepository} from "../../repository/user.repository";
+import {User} from "../../model/user";
+import {SessionPartial} from "../../model/session.partial";
+import {Session} from "../../model/session";
+import {SessionRepository} from "../../repository/session.repository";
 
 const userRepository = container.get<UserRepository>(TYPES.UserRepository);
 const sessionRepository = container.get<SessionRepository>(TYPES.SessionRepository);
