+++ express-backend-api/test/api/user.api.test.ts
@@ -1,12 +1,12 @@
 import request from "supertest";
-import {UserPartial} from "../model/user.partial";
+import {UserPartial} from "../../model/user.partial";
 import {faker} from "@faker-js/faker";
-import {container} from "../config/container.config";
-import {UserRepository} from "../repository/user.repository";
-import {TYPES} from "../config/types.config";
-import app from "../app";
-import {API_URI_LIR} from "../config/api.config";
-import {User} from "../model/user";
+import {container} from "../../config/container.config";
+import {UserRepository} from "../../repository/user.repository";
+import {TYPES} from "../../config/types.config";
+import app from "../../app";
+import {API_URI_LIR} from "../../config/api.config";
+import {User} from "../../model/user";
 import {randomBytes} from "crypto";
 import {ObjectId} from "mongodb";
